Feature: #140 예약 화면, 반납 화면 예외처리 Result로 감싸 전달, View에 표시

+++ domain/src/main/java/com/gta/domain/model/Reservation.kt
@@ -10,7 +10,7 @@ data class Reservation(
     val ownerId: String = "정보 없음",
     val state: Int = ReservationState.PENDING.state,
     val reservationDate: AvailableDate = AvailableDate(),
-    val price: Int = 0,
+    val price: Long = 0,
     val insuranceOption: String = InsuranceOption.LOW.name
 )
